api/requirements.txt

tensorflow
fastapi==0.105.0
keras==2.8.0
numpy==1.26.2
Pillow==10.1.0
pydantic==2.5.2
torch==2.1.1
